backend/generate-site/requirements.txt
psycopg2-binary==2.9.9
groq==0.13.0